Write "Hello World" test page.
It is not working. The latest error indicates that FitNesse cannot find
the runner class. My working hypothesis is that I have a typo in the
name of the runner or that fully qualified type names using relative
paths is incorrect.

Before changing too much, I wanted to snapshot my progress so far.